refactor: update HashCard and TitleCard
